Deps reorg. Subscriptions management WIP.
