+++ Cargo.toml
@@ -29,21 +29,22 @@ members = [
     "ipld/blockstore",
     # TODO re-enable graphsync when work continues
     # "ipld/graphsync",
-    "utils/bigint",
-    "tests/serialization_tests",
-    "utils/auth",
+    "key_management",
     "tests/conformance_tests",
-    "utils/bitfield",
-    "utils/test_utils",
-    "utils/commcid",
-    "utils/json_utils",
-    "utils/genesis",
-    "utils/net_utils",
-    "utils/statediff",
-    "utils/hash_utils",
+    "tests/serialization_tests",
     "types",
     "types/networks",
-    "key_management",
+    "types/rpc",
+    "utils/auth",
+    "utils/bigint",
+    "utils/bitfield",
+    "utils/commcid",
+    "utils/genesis",
+    "utils/hash_utils",
+    "utils/json_utils",
+    "utils/net_utils",
+    "utils/statediff",
+    "utils/test_utils",
 ]
 
 [patch.crates-io]
