+++ node/rpc/Cargo.toml
@@ -5,61 +5,63 @@ authors = ["ChainSafe Systems <info@chainsafe.io>"]
 edition = "2018"
 
 [dependencies]
-ipld_amt = "0.2"
-actor = { package = "actor_interface", path = "../../vm/actor_interface" }
+ahash = "0.6"
+async-log = "2.0.0"
+async-std = { version = "1.9", features = ["attributes"] }
+async-tungstenite = "0.12"
+base64 = "0.13"
 bls-signatures = { version = "0.7", default-features = false, features = [
     "blst"
 ] }
-async-std = { version = "1.9", features = ["attributes"] }
+crossbeam = "0.8.0"
+futures = "0.3.5"
+jsonwebtoken = "7.2.0"
+libp2p = { version = "0.28", default-features = false }
+log = "0.4.8"
+num-traits = "0.2.11"
+rand = "0.7"
+rand_distr = "0.3"
 serde = { version = "1.0.101", default-features = false, features = ["derive"] }
 serde_json = "1.0.48"
-chain = { path = "../../blockchain/chain", features = ["json"] }
-chain_sync = { path = "../../blockchain/chain_sync" }
-blockstore = { package = "ipld_blockstore", version = "0.1" }
-cid = { package = "forest_cid", version = "0.3", features = ["json"] }
-blocks = { package = "forest_blocks", path = "../../blockchain/blocks", features = [
-    "json"
-] }
+thiserror = "1.0"
+tide = "0.16.0"
+tide-websockets = "0.2.0"
+tokio = { version = "1.0", features = ["sync"] }
+##
+actor = { package = "actor_interface", path = "../../vm/actor_interface" }
+address = { package = "forest_address", version = "0.3", features = ["json"] }
+auth = { path = "../../utils/auth" }
 beacon = { package = "beacon", path = "../../blockchain/beacon", features = [
     "json"
 ] }
+bitfield = { package = "forest_bitfield", version = "0.1", features = ["json"] }
+blocks = { package = "forest_blocks", path = "../../blockchain/blocks", features = [
+    "json"
+] }
+blockstore = { package = "ipld_blockstore", version = "0.1" }
+chain = { path = "../../blockchain/chain", features = ["json"] }
+chain_sync = { path = "../../blockchain/chain_sync" }
+cid = { package = "forest_cid", version = "0.3", features = ["json"] }
 clock = { package = "fil_clock", path = "../clock" }
+crypto = { package = "forest_crypto", version = "0.4", features = ["json"] }
+encoding = { package = "forest_encoding", version = "0.2.1" }
+fil_types = "0.1"
+forest_libp2p = { path = "../forest_libp2p" }
+interpreter = { path = "../../vm/interpreter/" }
+ipld = { package = "forest_ipld", path = "../../ipld", features = ["json"] }
+ipld_amt = "0.2"
 message = { package = "forest_message", version = "0.6", features = ["json"] }
 message_pool = { path = "../../blockchain/message_pool" }
-crypto = { package = "forest_crypto", version = "0.4", features = ["json"] }
-num-traits = "0.2.11"
+networks = { path = "../../types/networks" }
+num-bigint = { version = "0.1.1", package = "forest_bigint" }
+rpc_types = { path = "../../types/rpc" }
+state_manager = { package = "state_manager", path = "../../blockchain/state_manager" }
+state_tree = { path = "../../vm/state_tree" }
+utils = { path = "../../node/utils" }
+vm = { package = "forest_vm", version = "0.3.1" }
 wallet = { package = "key_management", path = "../../key_management", features = [
     "json"
 ] }
-state_manager = { package = "state_manager", path = "../../blockchain/state_manager" }
-address = { package = "forest_address", version = "0.3", features = ["json"] }
-encoding = { package = "forest_encoding", version = "0.2.1" }
-num-bigint = { version = "0.1.1", package = "forest_bigint" }
-thiserror = "1.0"
-state_tree = { path = "../../vm/state_tree" }
-forest_libp2p = { path = "../forest_libp2p" }
-jsonwebtoken = "7.2.0"
-auth = { path = "../../utils/auth" }
-utils = { path = "../../node/utils" }
-rand_distr = "0.3"
-rand = "0.7"
-interpreter = { path = "../../vm/interpreter/" }
-fil_types = "0.1"
-futures = "0.3.5"
-async-tungstenite = "0.12"
-async-log = "2.0.0"
-log = "0.4.8"
-tokio = { version = "1.0", features = ["sync"] }
-libp2p = { version = "0.28", default-features = false }
-vm = { package = "forest_vm", version = "0.3.1" }
-base64 = "0.13"
-ipld = { package = "forest_ipld", path = "../../ipld", features = ["json"] }
-bitfield = { package = "forest_bitfield", version = "0.1", features = ["json"] }
-networks = { path = "../../types/networks" }
-ahash = "0.6"
-tide = "0.16.0"
-tide-websockets = "0.2.0"
-crossbeam = "0.8.0"
 
 [dependencies.jsonrpc-v2]
 version = "0.8.0"
@@ -69,7 +71,7 @@ default-features = false
 [dev-dependencies]
 db = { package = "forest_db", version = "0.1" }
 futures = "0.3.5"
+hex = "0.4.2"
 test_utils = { version = "0.1.0", path = "../../utils/test_utils/", features = [
     "test_constructors"
 ] }
-hex = "0.4.2"
