+++ types/rpc/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rpc_types"
+version = "0.1.0"
+authors = ["ChainSafe Systems <info@chainsafe.io>"]
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0.101", default-features = false, features = ["derive"] }
+serde_json = "1.0.48"
